% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afl-data.R
\docType{data}
\name{afl.finalists}
\alias{afl.finalists}
\title{AFL finalists}
\format{
A data frame with 400 rows and 1 variable:
\describe{
\item{afl.finalists}{Factor with 17 levels giving the team name of each
finalist appearance.}
}
}
\source{
Navarro, D. J. (2019). \emph{Learning statistics with R}.
}
\usage{
afl.finalists
}
\description{
A dataset listing the teams that appeared in the finals of the Australian
Football League (AFL) premiership from 1987 to 2010.
}
\keyword{datasets}
