% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afl-data.R
\docType{data}
\name{afl.margins}
\alias{afl.margins}
\title{AFL winning margins}
\format{
A data frame with 176 rows and 1 variable:
\describe{
\item{afl.margins}{Numeric. Winning margin in points.}
}
}
\source{
Navarro, D. J. (2019). \emph{Learning statistics with R}.
}
\usage{
afl.margins
}
\description{
The winning margin (in points) for each game played in the AFL (Australian
Football League) from 1987 to 2010 that was not a draw.
}
\keyword{datasets}
