% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afl-data.R
\docType{data}
\name{afl2}
\alias{afl2}
\title{AFL match margins and years}
\format{
A data frame with 4296 rows and 2 variables:
\describe{
\item{margin}{Numeric. Winning margin in points (negative values indicate
the away team won).}
\item{year}{Numeric. Year the game was played.}
}
}
\source{
Navarro, D. J. (2019). \emph{Learning statistics with R}.
}
\usage{
afl2
}
\description{
A simplified AFL dataset containing the winning margin and year for every
game played from 1987 to 2010.
}
\keyword{datasets}
