% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agpp-booksales-cakes.R
\docType{data}
\name{agpp}
\alias{agpp}
\title{Attitude to global poverty poll}
\format{
A data frame with 100 rows and 3 variables:
\describe{
\item{id}{Factor. Participant identifier.}
\item{response_before}{Factor with 2 levels (\code{"no"}, \code{"yes"}).
Response before the intervention.}
\item{response_after}{Factor with 2 levels (\code{"no"}, \code{"yes"}).
Response after the intervention.}
}
}
\source{
Navarro, D. J. (2019). \emph{Learning statistics with R}.
}
\usage{
agpp
}
\description{
A dataset recording whether 100 survey respondents answered "yes" or "no"
to a question about awareness of global poverty, measured on two occasions
(before and after an intervention).
}
\keyword{datasets}
