% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anscombe-awesome-baddata.R
\docType{data}
\name{awesome}
\alias{awesome}
\title{Scores by group (awesome)}
\format{
A data frame with 10 rows and 2 variables:
\describe{
\item{scores}{Numeric. Test score.}
\item{group}{Factor with 2 levels (\code{"A"}, \code{"B"}).}
}
}
\source{
Navarro, D. J. (2019). \emph{Learning statistics with R}.
}
\usage{
awesome
}
\description{
A small dataset of test scores for participants in two groups, used to
illustrate independent-samples comparisons.
}
\keyword{datasets}
