% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agpp-booksales-cakes.R
\docType{data}
\name{cakes}
\alias{cakes}
\title{Cakes matrix}
\format{
A numeric matrix with 4 rows and 5 columns.
}
\source{
Navarro, D. J. (2019). \emph{Learning statistics with R}.
}
\usage{
cakes
}
\description{
A matrix of test scores used to illustrate matrix operations and indexing
in R.
}
\keyword{datasets}
