% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cards-chapek9-chico-choice.R
\docType{data}
\name{cards}
\alias{cards}
\title{Card suit choices (study 1 and study 2)}
\format{
A data frame with 200 rows and 3 variables:
\describe{
\item{id}{Factor. Participant identifier.}
\item{choice_1}{Factor with 4 levels (\code{"clubs"}, \code{"diamonds"},
\code{"hearts"}, \code{"spades"}). First card suit chosen.}
\item{choice_2}{Factor with 4 levels (\code{"clubs"}, \code{"diamonds"},
\code{"hearts"}, \code{"spades"}). Second card suit chosen.}
}
}
\source{
Navarro, D. J. (2019). \emph{Learning statistics with R}.
}
\usage{
cards
}
\description{
Two hundred participants each made two choices of playing card suit.  Used
to illustrate chi-square tests of independence.
}
\keyword{datasets}
