% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cards-chapek9-chico-choice.R
\docType{data}
\name{chapek9}
\alias{chapek9}
\title{Chapek 9 robot/human data}
\format{
A data frame with 180 rows and 2 variables:
\describe{
\item{species}{Factor with 2 levels (\code{"robot"}, \code{"human"}).}
\item{choice}{Factor with 3 levels (\code{"puppy"}, \code{"flower"},
\code{"robot"}). The item chosen by the participant.}
}
}
\source{
Navarro, D. J. (2019). \emph{Learning statistics with R}.
}
\usage{
chapek9
}
\description{
Data from a hypothetical study on the planet Chapek 9, where humans and
robots are asked to choose between a puppy, a flower, or a robot.  Used to
illustrate chi-square goodness-of-fit and tests of independence.
}
\keyword{datasets}
