% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clinicaltrial-coffee.R
\docType{data}
\name{clinicaltrial}
\alias{clinicaltrial}
\title{Clinical trial data}
\format{
A data frame with 18 rows and 3 variables:
\describe{
\item{drug}{Factor with 3 levels (\code{"placebo"}, \code{"anxifree"},
\code{"joyzepam"}).}
\item{therapy}{Factor with 2 levels (\code{"no.therapy"}, \code{"CBT"}).}
\item{mood.gain}{Numeric. Change in mood score after treatment.}
}
}
\source{
Navarro, D. J. (2019). \emph{Learning statistics with R}.
}
\usage{
clinicaltrial
}
\description{
Results from a hypothetical clinical trial examining the effect of two
drugs (anxifree and joyzepam) and cognitive behavioural therapy (CBT) on
mood.  Used to illustrate factorial ANOVA.
}
\keyword{datasets}
