% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clinicaltrial-coffee.R
\docType{data}
\name{clinicaltrial_old}
\alias{clinicaltrial_old}
\title{Clinical trial data (older version)}
\format{
A data frame with 18 rows and 3 variables:
\describe{
\item{drug}{Factor with 3 levels (\code{"placebo"}, \code{"anxifree"},
\code{"joyzepam"}).}
\item{gender}{Factor with 2 levels (\code{"male"}, \code{"female"}).}
\item{mood.gain}{Numeric. Change in mood score after treatment.}
}
}
\source{
Navarro, D. J. (2019). \emph{Learning statistics with R}.
}
\usage{
clinicaltrial_old
}
\description{
An earlier version of the \code{\link{clinicaltrial}} dataset in which the
second grouping variable is \code{gender} rather than \code{therapy}.
}
\keyword{datasets}
