% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clinicaltrial-coffee.R
\docType{data}
\name{coffee}
\alias{coffee}
\title{Coffee preference and loquaciousness}
\format{
A data frame with 18 rows and 3 variables:
\describe{
\item{milk}{Factor with 2 levels (\code{"yes"}, \code{"no"}).
Whether the participant takes milk.}
\item{sugar}{Factor with 3 levels (\code{"none"}, \code{"fake"},
\code{"real"}). Type of sweetener used.}
\item{babble}{Numeric. Amount of babbling (higher = more talkative).}
}
}
\source{
Navarro, D. J. (2019). \emph{Learning statistics with R}.
}
\usage{
coffee
}
\description{
A dataset recording coffee drinkers' preference for milk and sugar together
with a measure of how much they babble (talk incessantly).  Used to
illustrate factorial ANOVA.
}
\keyword{datasets}
