% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drugs-dwr.R
\docType{data}
\name{drugs}
\alias{drugs}
\title{Drug effects on working memory and reaction time}
\format{
A data frame with 10 rows and 8 variables:
\describe{
\item{id}{Factor. Participant identifier.}
\item{gender}{Factor with 2 levels (\code{"male"}, \code{"female"}).}
\item{WMC_alcohol}{Numeric. Working memory capacity under alcohol.}
\item{WMC_caffeine}{Numeric. Working memory capacity under caffeine.}
\item{WMC_no.drug}{Numeric. Working memory capacity with no drug.}
\item{RT_alcohol}{Numeric. Reaction time (ms) under alcohol.}
\item{RT_caffeine}{Numeric. Reaction time (ms) under caffeine.}
\item{RT_no.drug}{Numeric. Reaction time (ms) with no drug.}
}
}
\source{
Navarro, D. J. (2019). \emph{Learning statistics with R}.
}
\usage{
drugs
}
\description{
Results from a hypothetical experiment in which 10 participants were tested
under three drug conditions (alcohol, caffeine, no drug).  Both working
memory capacity (WMC) and reaction time (RT) were recorded in each
condition.  This wide-format version is used to illustrate repeated-measures
ANOVA and related analyses.
}
\seealso{
\code{\link{dwr}}, \code{\link{drugs.wmc.rt}},
\code{\link{dwr.long}}, \code{\link{dwr.wide}}
}
\keyword{datasets}
