% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drugs-dwr.R
\docType{data}
\name{drugs.wmc.rt}
\alias{drugs.wmc.rt}
\title{Drug effects on working memory and reaction time (copy)}
\format{
A data frame with 10 rows and 8 variables. See
\code{\link{drugs}} for a full description of variables.
}
\source{
Navarro, D. J. (2019). \emph{Learning statistics with R}.
}
\usage{
drugs.wmc.rt
}
\description{
An alternative copy of the \code{\link{drugs}} dataset with the same
structure, used in different chapters of the textbook.
}
\seealso{
\code{\link{drugs}}, \code{\link{dwr}}
}
\keyword{datasets}
