% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drugs-dwr.R
\docType{data}
\name{dwr.mv}
\alias{dwr.mv}
\title{Drug/working memory/RT data (multivariate list)}
\format{
A list with 4 elements corresponding to the outcome matrices,
participant IDs, and gender coding used in multivariate analyses.
}
\source{
Navarro, D. J. (2019). \emph{Learning statistics with R}.
}
\usage{
dwr.mv
}
\description{
The \code{\link{dwr}} experiment data stored as a list, used to illustrate
multivariate analyses such as MANOVA.
}
\seealso{
\code{\link{dwr}}, \code{\link{dwr.long}}, \code{\link{dwr.wide}}
}
\keyword{datasets}
