% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drugs-dwr.R
\docType{data}
\name{dwr}
\alias{dwr}
\title{Drug/working memory/RT data (wide format, original)}
\format{
A data frame with 10 rows and 8 variables. See
\code{\link{drugs}} for a full description of variables.
}
\source{
Navarro, D. J. (2019). \emph{Learning statistics with R}.
}
\usage{
dwr
}
\description{
The same drug-effects experiment as \code{\link{drugs}}, stored here under
the name \code{dwr}.
}
\seealso{
\code{\link{drugs}}, \code{\link{dwr.long}}, \code{\link{dwr.wide}}
}
\keyword{datasets}
