% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drugs-dwr.R
\docType{data}
\name{dwr.wide}
\alias{dwr.wide}
\title{Drug/working memory/RT data (fully wide format)}
\format{
A data frame with 10 rows and 14 variables:
\describe{
\item{id}{Factor. Participant identifier.}
\item{gender}{Factor with 2 levels (\code{"male"}, \code{"female"}).}
\item{wm_alc_t1}{Numeric. Working memory, alcohol, time 1.}
\item{wm_caf_t1}{Numeric. Working memory, caffeine, time 1.}
\item{wm_nil_t1}{Numeric. Working memory, no drug, time 1.}
\item{wm_alc_t2}{Numeric. Working memory, alcohol, time 2.}
\item{wm_caf_t2}{Numeric. Working memory, caffeine, time 2.}
\item{wm_nil_t2}{Numeric. Working memory, no drug, time 2.}
\item{rt_alc_t1}{Numeric. Reaction time (ms), alcohol, time 1.}
\item{rt_caf_t1}{Numeric. Reaction time (ms), caffeine, time 1.}
\item{rt_nil_t1}{Numeric. Reaction time (ms), no drug, time 1.}
\item{rt_alc_t2}{Numeric. Reaction time (ms), alcohol, time 2.}
\item{rt_caf_t2}{Numeric. Reaction time (ms), caffeine, time 2.}
\item{rt_nil_t2}{Numeric. Reaction time (ms), no drug, time 2.}
}
}
\source{
Navarro, D. J. (2019). \emph{Learning statistics with R}.
}
\usage{
dwr.wide
}
\description{
The \code{\link{dwr}} experiment data in a wider format that separates both
drug condition and measurement occasion into separate columns.
}
\seealso{
\code{\link{dwr}}, \code{\link{dwr.long}}
}
\keyword{datasets}
