% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effort-garden-grades.R
\docType{data}
\name{garden}
\alias{garden}
\title{Night Garden utterances with context}
\format{
A data frame with 5 rows and 3 variables:
\describe{
\item{speaker}{Factor with 3 levels (\code{"makka-pakka"},
\code{"tombliboo"}, \code{"upsy-daisy"}).  Character who spoke.}
\item{utterance}{Factor with 4 levels (\code{"ee"}, \code{"onk"},
\code{"oo"}, \code{"pip"}).  Word or sound uttered.}
\item{line}{Numeric. Line number in the script.}
}
}
\source{
Navarro, D. J. (2019). \emph{Learning statistics with R}.
}
\usage{
garden
}
\description{
A small dataset recording utterances made by characters from the children's
television show \emph{In the Night Garden}, along with the line number in
the script.  Used to illustrate data frames and basic data manipulation.
}
\seealso{
\code{\link{speaker}}, \code{\link{utterance}}
}
\keyword{datasets}
