% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effort-garden-grades.R
\docType{data}
\name{grade.A}
\alias{grade.A}
\title{Exam grades for class A}
\format{
A data frame with 15 rows and 1 variable:
\describe{
\item{grade.A}{Numeric. Exam grade (percentage).}
}
}
\source{
Navarro, D. J. (2019). \emph{Learning statistics with R}.
}
\usage{
grade.A
}
\description{
Exam grades for 15 students in class A.  Used alongside
\code{\link{grade.B}} to illustrate two-sample tests.
}
\seealso{
\code{\link{grade.B}}, \code{\link{harpo}}
}
\keyword{datasets}
