% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effort-garden-grades.R
\docType{data}
\name{grade.B}
\alias{grade.B}
\title{Exam grades for class B}
\format{
A data frame with 18 rows and 1 variable:
\describe{
\item{grade.B}{Numeric. Exam grade (percentage).}
}
}
\source{
Navarro, D. J. (2019). \emph{Learning statistics with R}.
}
\usage{
grade.B
}
\description{
Exam grades for 18 students in class B.  Used alongside
\code{\link{grade.A}} to illustrate two-sample tests.
}
\seealso{
\code{\link{grade.A}}, \code{\link{harpo}}
}
\keyword{datasets}
