% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effort-garden-grades.R
\docType{data}
\name{grades}
\alias{grades}
\title{Student grades}
\format{
A data frame with 20 rows and 1 variable:
\describe{
\item{grades}{Numeric. Exam grade (percentage).}
}
}
\source{
Navarro, D. J. (2019). \emph{Learning statistics with R}.
}
\usage{
grades
}
\description{
Exam grades for 20 students.  Used to illustrate descriptive statistics and
frequency distributions.
}
\keyword{datasets}
