% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/happiness-harpo-likert.R
\docType{data}
\name{harpo}
\alias{harpo}
\title{Student grades by tutor (Harpo)}
\format{
A data frame with 33 rows and 2 variables:
\describe{
\item{grade}{Numeric. Exam grade (percentage).}
\item{tutor}{Factor with 2 levels (\code{"Anastasia"},
\code{"Bernadette"}).  Tutor assigned to the student.}
}
}
\source{
Navarro, D. J. (2019). \emph{Learning statistics with R}.
}
\usage{
harpo
}
\description{
Exam grades for 33 students assigned to one of two tutors (Anastasia or
Bernadette).  Used to illustrate the independent-samples t-test.
}
\seealso{
\code{\link{grade.A}}, \code{\link{grade.B}}
}
\keyword{datasets}
