% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parenthood.R
\docType{data}
\name{parenthood2}
\alias{parenthood2}
\title{Parenthood data with missing values}
\format{
A data frame with 100 rows and 4 variables. See
\code{\link{parenthood}} for a description of the variables.
}
\source{
Navarro, D. J. (2019). \emph{Learning statistics with R}.
}
\usage{
parenthood2
}
\description{
A version of \code{\link{parenthood}} with some values set to \code{NA},
used to illustrate handling of missing data.
}
\seealso{
\code{\link{parenthood}}, \code{\link{parenthood3}}
}
\keyword{datasets}
