% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drugs-dwr.R
\docType{data}
\name{rms}
\alias{rms}
\title{Repeated-measures specification list}
\format{
A list with 2 elements: a named vector describing the wide-format
column layout and a character vector of between-subjects variable names.
}
\source{
Navarro, D. J. (2019). \emph{Learning statistics with R}.
}
\usage{
rms
}
\description{
A helper list used internally to specify the repeated-measures structure of
the \code{\link{dwr.wide}} dataset (variable mapping, measure indices,
drug levels, and time levels).
}
\seealso{
\code{\link{dwr.wide}}
}
\keyword{datasets}
