% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtfm-scores-suits-trial.R
\docType{data}
\name{rtfm.1}
\alias{rtfm.1}
\title{Student grades, attendance and reading (numeric)}
\format{
A data frame with 8 rows and 3 variables:
\describe{
\item{grade}{Numeric. Final course grade (percentage).}
\item{attend}{Numeric. Whether the student attended class (\code{1} = yes,
\code{0} = no).}
\item{reading}{Numeric. Whether the student did the readings (\code{1} =
yes, \code{0} = no).}
}
}
\source{
Navarro, D. J. (2019). \emph{Learning statistics with R}.
}
\usage{
rtfm.1
}
\description{
Grades and binary indicators of class attendance and textbook reading for
8 students.  Used to illustrate multiple regression with binary predictors.
}
\seealso{
\code{\link{rtfm.2}}
}
\keyword{datasets}
