% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtfm-scores-suits-trial.R
\docType{data}
\name{score.A}
\alias{score.A}
\title{Scores for group A (small sample)}
\format{
A data frame with 5 rows and 1 variable:
\describe{
\item{score.A}{Numeric. Test score.}
}
}
\source{
Navarro, D. J. (2019). \emph{Learning statistics with R}.
}
\usage{
score.A
}
\description{
The five scores from group A of the \code{\link{awesome}} dataset, stored
separately.  Used to illustrate basic vector operations.
}
\seealso{
\code{\link{score.B}}, \code{\link{awesome}}
}
\keyword{datasets}
