% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtfm-scores-suits-trial.R
\docType{data}
\name{suit.choice}
\alias{suit.choice}
\title{Card suit choices (marginal)}
\format{
A data frame with 200 rows and 1 variable:
\describe{
\item{suit.choice}{Factor with 4 levels (\code{"clubs"},
\code{"diamonds"}, \code{"hearts"}, \code{"spades"}).}
}
}
\source{
Navarro, D. J. (2019). \emph{Learning statistics with R}.
}
\usage{
suit.choice
}
\description{
The first card suit choice made by each of 200 participants.  Used to
illustrate chi-square goodness-of-fit tests.
}
\seealso{
\code{\link{cards}}
}
\keyword{datasets}
