% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speaker-utterance-work.R
\docType{data}
\name{work2}
\alias{work2}
\title{Work productivity data with missing values}
\format{
A data frame with 49 rows and 7 variables. See \code{\link{work}}
for a description of all variables.
}
\source{
Navarro, D. J. (2019). \emph{Learning statistics with R}.
}
\usage{
work2
}
\description{
A version of \code{\link{work}} with some values in \code{tasks} and
\code{pay} set to \code{NA}, used to illustrate missing data handling.
}
\seealso{
\code{\link{work}}
}
\keyword{datasets}
