% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{afl.finalists}
\alias{afl.finalists}
\title{AFL Finals Teams (2000--2010)}
\format{
A data frame with 400 rows and 1 variable:
\describe{
  \item{afl.finalists}{Name of the team (factor with 16 levels)}
}
}
\usage{
afl.finalists
}
\description{
A data frame recording which AFL (Australian Football League) teams made it
to the finals for each year from 2000 to 2010.
}
\source{
Navarro, D. J. (2019). \emph{Learning statistics with R: A tutorial for
psychology students and other beginners} (version 0.6.1).
\url{https://learningstatisticswithr.com/}
}
\keyword{datasets}
