% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{afl.margins}
\alias{afl.margins}
\title{AFL Winning Margins (2010)}
\format{
A data frame with 176 rows and 1 variable:
\describe{
  \item{afl.margins}{Winning margin in points (numeric)}
}
}
\usage{
afl.margins
}
\description{
A data frame of winning margins from 176 AFL (Australian Football League)
games played in the 2010 regular season.
}
\source{
Navarro, D. J. (2019). \emph{Learning statistics with R: A tutorial for
psychology students and other beginners} (version 0.6.1).
\url{https://learningstatisticswithr.com/}
}
\keyword{datasets}
