% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{afl}
\alias{afl}
\title{AFL Game Records (1987--2010)}
\format{
A data frame with 4296 rows and 12 variables:
\describe{
  \item{home.team}{Name of the home team (factor)}
  \item{away.team}{Name of the away team (factor)}
  \item{home.score}{Score of the home team (numeric)}
  \item{away.score}{Score of the away team (numeric)}
  \item{year}{Year the game was played (numeric)}
  \item{round}{Round number (numeric)}
  \item{weekday}{Day of the week (factor)}
  \item{day}{Day of the month (numeric)}
  \item{month}{Month number (numeric)}
  \item{is.final}{Whether the game was a final (logical)}
  \item{venue}{Venue where the game was played (factor)}
  \item{attendance}{Number of spectators (numeric)}
}
}
\usage{
afl
}
\description{
A data frame containing records from 4296 AFL (Australian Football League)
games played between 1987 and 2010.
}
\source{
Navarro, D. J. (2019). \emph{Learning statistics with R: A tutorial for
psychology students and other beginners} (version 0.6.1).
\url{https://learningstatisticswithr.com/}
}
\keyword{datasets}
