% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{anscombesquartet}
\alias{anscombesquartet}
\title{Anscombe's Quartet}
\format{
A data frame with 11 rows and 8 variables:
\describe{
  \item{X1}{x-values for dataset 1 (numeric)}
  \item{X2}{x-values for dataset 2 (numeric)}
  \item{X3}{x-values for dataset 3 (numeric)}
  \item{X4}{x-values for dataset 4 (numeric)}
  \item{Y1}{y-values for dataset 1 (numeric)}
  \item{Y2}{y-values for dataset 2 (numeric)}
  \item{Y3}{y-values for dataset 3 (numeric)}
  \item{Y4}{y-values for dataset 4 (numeric)}
}
}
\usage{
anscombesquartet
}
\description{
A data frame containing Anscombe's Quartet, a set of four datasets that
have nearly identical simple descriptive statistics yet differ significantly
when graphed.
}
\source{
Anscombe, F. J. (1973). Graphs in statistical analysis. \emph{The
American Statistician}, 27(1), 17--21.
\doi{10.1080/00031305.1973.10478966}
}
\keyword{datasets}
