% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{awesome}
\alias{awesome}
\title{Awesome Scores}
\format{
A data frame with 10 rows and 2 variables:
\describe{
  \item{scores}{Score for each participant (numeric)}
  \item{group}{Group assignment (factor: "A", "B")}
}
}
\usage{
awesome
}
\description{
A small toy data frame used to illustrate the independent samples t-test.
Contains scores for 10 participants assigned to one of two groups.
}
\source{
Navarro, D. J. (2019). \emph{Learning statistics with R: A tutorial for
psychology students and other beginners} (version 0.6.1).
\url{https://learningstatisticswithr.com/}
}
\keyword{datasets}
