% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{booksales}
\alias{booksales}
\title{Monthly Book Sales}
\format{
A data frame with 12 rows and 4 variables:
\describe{
  \item{Month}{Month name (character)}
  \item{Days}{Number of days in the month (integer)}
  \item{Sales}{Number of books sold (integer)}
  \item{Stock.Levels}{Inventory status (character)}
}
}
\usage{
booksales
}
\description{
A data frame recording monthly book sales data over one year, used to
illustrate data import and manipulation with data frames.
}
\source{
Navarro, D. J. (2019). \emph{Learning statistics with R: A tutorial for
psychology students and other beginners} (version 0.6.1).
\url{https://learningstatisticswithr.com/}
}
\keyword{datasets}
