% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{cards}
\alias{cards}
\title{Card Suit Choices}
\format{
A data frame with 200 rows and 3 variables:
\describe{
  \item{id}{Participant identifier (factor)}
  \item{choice_1}{Suit chosen on the first draw (factor: "clubs",
    "diamonds", "hearts", "spades")}
  \item{choice_2}{Suit chosen on the second draw (factor: "clubs",
    "diamonds", "hearts", "spades")}
}
}
\usage{
cards
}
\description{
A data frame recording the card suit choices made by 200 participants in
two successive draws, used to illustrate tests of association.
}
\source{
Navarro, D. J. (2019). \emph{Learning statistics with R: A tutorial for
psychology students and other beginners} (version 0.6.1).
\url{https://learningstatisticswithr.com/}
}
\keyword{datasets}
