% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{chapek9}
\alias{chapek9}
\title{Chapek 9 Robot Detection Study}
\format{
A data frame with 180 rows and 2 variables:
\describe{
  \item{species}{Species of the participant (factor: "human", "robot")}
  \item{choice}{Object chosen by the participant (factor: "data",
    "flower", "puppy")}
}
}
\usage{
chapek9
}
\description{
A data frame from a hypothetical study in which 180 participants (robots
and humans) were asked to choose between three objects (a bunch of flowers,
a data set, or a puppy). The study is named after Chapek 9, a fictional
planet from the animated television series \emph{Futurama}, and is used to
illustrate the chi-square test of independence.
}
\source{
Navarro, D. J. (2019). \emph{Learning statistics with R: A tutorial for
psychology students and other beginners} (version 0.6.1).
\url{https://learningstatisticswithr.com/}
}
\keyword{datasets}
