% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{chico}
\alias{chico}
\title{Dr Chico's Student Grades}
\format{
A data frame with 20 rows and 3 variables:
\describe{
  \item{id}{Student identifier (factor)}
  \item{grade_test1}{Score on the first test (numeric)}
  \item{grade_test2}{Score on the second test (numeric)}
}
}
\usage{
chico
}
\description{
A data frame containing the test scores of 20 students from two tests
administered in Dr Chico's class, used to illustrate the paired-samples
t-test.
}
\source{
Navarro, D. J. (2019). \emph{Learning statistics with R: A tutorial for
psychology students and other beginners} (version 0.6.1).
\url{https://learningstatisticswithr.com/}
}
\keyword{datasets}
