% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{clinicaltrial}
\alias{clinicaltrial}
\title{Clinical Trial}
\format{
A data frame with 18 rows and 3 variables:
\describe{
  \item{drug}{Drug condition (factor: "anxifree", "joyzepam", "placebo")}
  \item{therapy}{Therapy condition (factor: "CBT", "no.therapy")}
  \item{mood.gain}{Improvement in mood score (numeric)}
}
}
\usage{
clinicaltrial
}
\description{
A data frame from a hypothetical clinical trial examining the effect of
two antidepressant drugs (anxifree and joyzepam) and cognitive-behavioural
therapy (CBT) on mood gain. Used to illustrate factorial ANOVA.
}
\source{
Navarro, D. J. (2019). \emph{Learning statistics with R: A tutorial for
psychology students and other beginners} (version 0.6.1).
\url{https://learningstatisticswithr.com/}
}
\keyword{datasets}
