% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{clinicaltrial_old}
\alias{clinicaltrial_old}
\title{Clinical Trial (Older Version)}
\format{
A data frame with 18 rows and 3 variables:
\describe{
  \item{drug}{Drug condition (factor: "anxifree", "joyzepam", "placebo")}
  \item{gender}{Gender of the participant (factor: "female", "male")}
  \item{mood.gain}{Improvement in mood score (numeric)}
}
}
\usage{
clinicaltrial_old
}
\description{
An older version of the \code{clinicaltrial} dataset in which the therapy
variable is replaced by a gender variable. Used to illustrate one-way ANOVA
and related analyses.
}
\source{
Navarro, D. J. (2019). \emph{Learning statistics with R: A tutorial for
psychology students and other beginners} (version 0.6.1).
\url{https://learningstatisticswithr.com/}
}
\keyword{datasets}
