% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{coffee}
\alias{coffee}
\title{Coffee and Babbling Experiment}
\format{
A data frame with 18 rows and 3 variables:
\describe{
  \item{milk}{Whether milk was added (factor: "no", "yes")}
  \item{sugar}{Type of sugar added (factor: "fake", "none", "real")}
  \item{babble}{Amount of incoherent babbling (numeric)}
}
}
\usage{
coffee
}
\description{
A data frame from a hypothetical experiment examining whether adding milk
and/or sugar to coffee affects the amount of incoherent babbling produced
by the drinker.
}
\source{
Navarro, D. J. (2019). \emph{Learning statistics with R: A tutorial for
psychology students and other beginners} (version 0.6.1).
\url{https://learningstatisticswithr.com/}
}
\keyword{datasets}
