% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{dwr.mv}
\alias{dwr.mv}
\title{Drug Effects on Working Memory and Reaction Time (Multivariate)}
\format{
A list with 4 elements:
\describe{
  \item{rt}{A 10x6 matrix of reaction time (ms) measurements, one column
    per drug-by-time combination}
  \item{wm}{A 10x6 matrix of working memory capacity scores, one column
    per drug-by-time combination}
  \item{id}{Participant identifier (factor)}
  \item{gender}{Gender of the participant (factor: "female", "male")}
}
}
\usage{
dwr.mv
}
\description{
A list containing multivariate data from a hypothetical experiment
examining the effects of alcohol, caffeine, and no drug on working memory
capacity and reaction time measured at two time points in 10 participants.
}
\source{
Navarro, D. J. (2019). \emph{Learning statistics with R: A tutorial for
psychology students and other beginners} (version 0.6.1).
\url{https://learningstatisticswithr.com/}
}
\keyword{datasets}
