% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{dwr.wide}
\alias{dwr.wide}
\title{Drug Effects on Working Memory and Reaction Time (Extended Wide Format)}
\format{
A data frame with 10 rows and 14 variables:
\describe{
  \item{id}{Participant identifier (factor)}
  \item{gender}{Gender of the participant (factor: "female", "male")}
  \item{wm_alc_t1}{Working memory capacity, alcohol condition, time 1 (numeric)}
  \item{wm_caf_t1}{Working memory capacity, caffeine condition, time 1 (numeric)}
  \item{wm_nil_t1}{Working memory capacity, no-drug condition, time 1 (numeric)}
  \item{wm_alc_t2}{Working memory capacity, alcohol condition, time 2 (numeric)}
  \item{wm_caf_t2}{Working memory capacity, caffeine condition, time 2 (numeric)}
  \item{wm_nil_t2}{Working memory capacity, no-drug condition, time 2 (numeric)}
  \item{rt_alc_t1}{Reaction time (ms), alcohol condition, time 1 (numeric)}
  \item{rt_caf_t1}{Reaction time (ms), caffeine condition, time 1 (numeric)}
  \item{rt_nil_t1}{Reaction time (ms), no-drug condition, time 1 (numeric)}
  \item{rt_alc_t2}{Reaction time (ms), alcohol condition, time 2 (numeric)}
  \item{rt_caf_t2}{Reaction time (ms), caffeine condition, time 2 (numeric)}
  \item{rt_nil_t2}{Reaction time (ms), no-drug condition, time 2 (numeric)}
}
}
\usage{
dwr.wide
}
\description{
A data frame from a hypothetical experiment examining the effects of
alcohol, caffeine, and no drug on working memory capacity and reaction
time measured at two time points in 10 participants.
}
\source{
Navarro, D. J. (2019). \emph{Learning statistics with R: A tutorial for
psychology students and other beginners} (version 0.6.1).
\url{https://learningstatisticswithr.com/}
}
\keyword{datasets}
