% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{effort}
\alias{effort}
\title{Effort and Exam Grade}
\format{
A data frame with 10 rows and 2 variables:
\describe{
  \item{hours}{Number of hours studied (numeric)}
  \item{grade}{Exam grade (numeric)}
}
}
\usage{
effort
}
\description{
A data frame recording study hours and exam grades for 10 students, used
to illustrate simple linear regression.
}
\source{
Navarro, D. J. (2019). \emph{Learning statistics with R: A tutorial for
psychology students and other beginners} (version 0.6.1).
\url{https://learningstatisticswithr.com/}
}
\keyword{datasets}
