% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{garden}
\alias{garden}
\title{In the Night Garden Utterances}
\format{
A data frame with 5 rows and 3 variables:
\describe{
  \item{speaker}{Character name (factor: "makka-pakka", "tombliboo",
    "upsy-daisy")}
  \item{utterance}{Word spoken (factor: "ee", "onk", "pip")}
  \item{line}{Line number (numeric)}
}
}
\usage{
garden
}
\description{
A data frame recording utterances made by characters from the British
children\'s television programme \emph{In the Night Garden}, used to
illustrate contingency tables and chi-square tests.
}
\source{
Navarro, D. J. (2019). \emph{Learning statistics with R: A tutorial for
psychology students and other beginners} (version 0.6.1).
\url{https://learningstatisticswithr.com/}
}
\keyword{datasets}
