% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{grade.A}
\alias{grade.A}
\title{Grades from Class A}
\format{
A data frame with 15 rows and 1 variable:
\describe{
  \item{grade.A}{Exam grade (numeric)}
}
}
\usage{
grade.A
}
\description{
A data frame containing exam grades from 15 students in class A, used to
illustrate descriptive statistics and hypothesis tests.
}
\source{
Navarro, D. J. (2019). \emph{Learning statistics with R: A tutorial for
psychology students and other beginners} (version 0.6.1).
\url{https://learningstatisticswithr.com/}
}
\keyword{datasets}
