% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{grade.B}
\alias{grade.B}
\title{Grades from Class B}
\format{
A data frame with 18 rows and 1 variable:
\describe{
  \item{grade.B}{Exam grade (numeric)}
}
}
\usage{
grade.B
}
\description{
A data frame containing exam grades from 18 students in class B, used to
illustrate descriptive statistics and hypothesis tests.
}
\source{
Navarro, D. J. (2019). \emph{Learning statistics with R: A tutorial for
psychology students and other beginners} (version 0.6.1).
\url{https://learningstatisticswithr.com/}
}
\keyword{datasets}
