% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{grades}
\alias{grades}
\title{Student Grades}
\format{
A data frame with 20 rows and 1 variable:
\describe{
  \item{grades}{Student grade (numeric)}
}
}
\usage{
grades
}
\description{
A data frame containing grades for 20 students, used to illustrate
one-sample hypothesis tests and descriptive statistics.
}
\source{
Navarro, D. J. (2019). \emph{Learning statistics with R: A tutorial for
psychology students and other beginners} (version 0.6.1).
\url{https://learningstatisticswithr.com/}
}
\keyword{datasets}
