% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{harpo}
\alias{harpo}
\title{Harpo's Students' Grades}
\format{
A data frame with 33 rows and 2 variables:
\describe{
  \item{grade}{Exam grade (numeric)}
  \item{tutor}{Tutor name (factor: "Anastasia", "Bernadette")}
}
}
\usage{
harpo
}
\description{
A data frame recording exam grades of 33 students taught by one of two
tutors (Anastasia or Bernadette), used to illustrate the independent
samples t-test.
}
\source{
Navarro, D. J. (2019). \emph{Learning statistics with R: A tutorial for
psychology students and other beginners} (version 0.6.1).
\url{https://learningstatisticswithr.com/}
}
\keyword{datasets}
