% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{likert.raw}
\alias{likert.raw}
\title{Raw Likert Scale Responses}
\format{
A data frame with 10 rows and 1 variable:
\describe{
  \item{likert.raw}{Likert scale rating from 1 to 7 (numeric)}
}
}
\usage{
likert.raw
}
\description{
A data frame containing raw responses on a 7-point Likert scale from
10 participants, used to illustrate analyses of ordinal data.
}
\source{
Navarro, D. J. (2019). \emph{Learning statistics with R: A tutorial for
psychology students and other beginners} (version 0.6.1).
\url{https://learningstatisticswithr.com/}
}
\keyword{datasets}
