% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{parenthood}
\alias{parenthood}
\title{Parenthood (Complete Data)}
\format{
A data frame with 100 rows and 4 variables:
\describe{
  \item{dan.sleep}{Hours of sleep Dan got (numeric)}
  \item{baby.sleep}{Hours of sleep the baby got (numeric)}
  \item{dan.grump}{Dan's grumpiness rating (0--100) (numeric)}
  \item{day}{Day number (integer)}
}
}
\usage{
parenthood
}
\description{
A data frame recording sleep and grumpiness data collected over 100 days,
used to illustrate correlation and linear regression. The data tracks how
much sleep Dan and his baby got each night and Dan's grumpiness the next
day.
}
\source{
Navarro, D. J. (2019). \emph{Learning statistics with R: A tutorial for
psychology students and other beginners} (version 0.6.1).
\url{https://learningstatisticswithr.com/}
}
\keyword{datasets}
