% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{parenthood3}
\alias{parenthood3}
\title{Parenthood (Extended Version)}
\format{
A data frame with 100 rows and 5 variables:
\describe{
  \item{dan.sleep}{Hours of sleep Dan got (numeric)}
  \item{baby.sleep}{Hours of sleep the baby got (numeric)}
  \item{dan.grump}{Dan's grumpiness rating (0--100) (numeric)}
  \item{day}{Day number (integer)}
  \item{weekday}{Day of the week (factor)}
}
}
\usage{
parenthood3
}
\description{
An extended version of the \code{parenthood} dataset that includes a
variable recording the day of the week, used to illustrate more complex
regression models.
}
\source{
Navarro, D. J. (2019). \emph{Learning statistics with R: A tutorial for
psychology students and other beginners} (version 0.6.1).
\url{https://learningstatisticswithr.com/}
}
\keyword{datasets}
