% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{rms}
\alias{rms}
\title{Repeated Measures Specification for Drug Study}
\format{
A list with 2 elements:
\describe{
  \item{within}{A data frame describing within-subject variables: the
    column name in the wide-format data, the measure name, the drug
    condition, and the time point}
  \item{between}{A character vector naming the between-subject variables
    (\code{"id"} and \code{"gender"})}
}
}
\usage{
rms
}
\description{
A list specifying the within-subject and between-subject structure of the
\code{dwr} repeated-measures experiment. Used as an argument to
repeated-measures ANOVA functions.
}
\source{
Navarro, D. J. (2019). \emph{Learning statistics with R: A tutorial for
psychology students and other beginners} (version 0.6.1).
\url{https://learningstatisticswithr.com/}
}
\keyword{datasets}
