% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{score.B}
\alias{score.B}
\title{Scores from Group B}
\format{
A data frame with 5 rows and 1 variable:
\describe{
  \item{score.B}{Test score (numeric)}
}
}
\usage{
score.B
}
\description{
A data frame containing 5 test scores from group B, used to illustrate
basic descriptive statistics.
}
\source{
Navarro, D. J. (2019). \emph{Learning statistics with R: A tutorial for
psychology students and other beginners} (version 0.6.1).
\url{https://learningstatisticswithr.com/}
}
\keyword{datasets}
