% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{speaker}
\alias{speaker}
\title{Speaker Names}
\format{
A data frame with 10 rows and 1 variable:
\describe{
  \item{speaker}{Character name of the speaker (character)}
}
}
\usage{
speaker
}
\description{
A data frame listing the names of 10 speakers, used alongside the
\code{utterance} dataset to illustrate how to work with character data.
}
\source{
Navarro, D. J. (2019). \emph{Learning statistics with R: A tutorial for
psychology students and other beginners} (version 0.6.1).
\url{https://learningstatisticswithr.com/}
}
\keyword{datasets}
