% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{suit.choice}
\alias{suit.choice}
\title{Card Suit Preferences}
\format{
A data frame with 200 rows and 1 variable:
\describe{
  \item{suit.choice}{Card suit chosen (factor: "clubs", "diamonds",
    "hearts", "spades")}
}
}
\usage{
suit.choice
}
\description{
A data frame recording the card suit chosen by 200 participants in a
hypothetical study, used to illustrate the goodness-of-fit chi-square test.
}
\source{
Navarro, D. J. (2019). \emph{Learning statistics with R: A tutorial for
psychology students and other beginners} (version 0.6.1).
\url{https://learningstatisticswithr.com/}
}
\keyword{datasets}
