% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{trial}
\alias{trial}
\title{Trial Data}
\format{
A data frame with 16 rows and 2 variables:
\describe{
  \item{happy}{Whether the participant was happy (logical)}
  \item{on.fire}{Whether the participant was on fire (logical)}
}
}
\usage{
trial
}
\description{
A data frame containing logical indicators for whether participants were
happy and/or on fire during a hypothetical trial, used to illustrate
basic probability concepts.
}
\source{
Navarro, D. J. (2019). \emph{Learning statistics with R: A tutorial for
psychology students and other beginners} (version 0.6.1).
\url{https://learningstatisticswithr.com/}
}
\keyword{datasets}
