% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{utterance}
\alias{utterance}
\title{Utterances}
\format{
A data frame with 10 rows and 1 variable:
\describe{
  \item{utterance}{The word or phrase spoken (character)}
}
}
\usage{
utterance
}
\description{
A data frame listing 10 utterances, used alongside the \code{speaker}
dataset to illustrate how to work with character data.
}
\source{
Navarro, D. J. (2019). \emph{Learning statistics with R: A tutorial for
psychology students and other beginners} (version 0.6.1).
\url{https://learningstatisticswithr.com/}
}
\keyword{datasets}
