% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{work}
\alias{work}
\title{Work Productivity Data (Complete)}
\format{
A data frame with 49 rows and 7 variables:
\describe{
  \item{hours}{Number of hours worked (numeric)}
  \item{tasks}{Number of tasks completed (numeric)}
  \item{pay}{Daily pay in dollars (numeric)}
  \item{day}{Day number (integer)}
  \item{weekday}{Day of the week (factor)}
  \item{week}{Week number (numeric)}
  \item{day.type}{Whether it is a weekday or weekend (factor: "weekday",
    "weekend")}
}
}
\usage{
work
}
\description{
A data frame recording work productivity information for 49 days over
7 weeks, used to illustrate regression with multiple predictors including
categorical variables.
}
\source{
Navarro, D. J. (2019). \emph{Learning statistics with R: A tutorial for
psychology students and other beginners} (version 0.6.1).
\url{https://learningstatisticswithr.com/}
}
\keyword{datasets}
